added format-price file and imported it into product-list-item and product-details
