+++ client/components/format-price.js
@@ -1,0 +1,8 @@
+function formatPrice(price) {
+  let formattedPrice = price + '';
+  formattedPrice = formattedPrice.split('');
+  formattedPrice.splice(parseInt(formattedPrice.length, 10) - 2, 0, '.');
+  return `$${formattedPrice.join('')}`;
+}
+
+export default formatPrice;
